make Windows builds work again
Fixes and VS2015 solultion/project files for 32- and 64-bit builds.
Older VS files have been removed. Adds a FIX::sys_socket_t type
surrogate for socket handles (unsigned on Windows).

+++ src/C++/Field.h
@@ -266,7 +266,7 @@ private:
   {
     if (!m_calculated)
     {
-      m_length = String::size(m_string) ;
+      m_length = (int)String::size(m_string) ;
       m_total = m_tagChecksum +
         Util::CharBuffer::checkSum(String::data(m_string), m_length) +
         (int)'\001';
